Guard bbox-init embed behind bbox_full build tag
The bbox-init guest binary is built by `task build-init` and is
.gitignore'd, so the Go module proxy serves the initbin package without
it. Embedding it unconditionally made brood-box uncompilable for any
downstream consumer importing it as a library (pkg/runtime -> pkg/sandbox
-> internal/infra/vm -> initbin):

  pattern bbox-init: no matching files found

Split the embed behind a `bbox_full` build tag (Option A from the issue):
embed_full.go carries the real //go:embed under the tag, embed_stub.go
exports an empty Binary by default. The bbox CLI build and install
targets now pass `bbox_full`; every other build (go test, golangci-lint,
library consumers) uses the stub and compiles without the file present.

The rootfs hook validates Binary is non-empty before use, so a stub
build that attempts to boot a VM fails fast with an actionable error
rather than writing a zero-byte init. Init-binary hook tests are split
by build tag accordingly.

Fixes #110
