+++ CLAUDE.md
@@ -226,6 +226,7 @@ Image tagging is `:latest` only — images are not versioned with release tags. 
 
 - **go-microvm is a tagged dependency**: `go.mod` depends on `github.com/stacklok/go-microvm` as a versioned module. `build` downloads pre-built go-microvm runtime artifacts and embeds them — no local checkout or system libkrun needed. Use `build-dev-system` to build go-microvm-runner from source (requires `libkrun-devel`).
 - **CGO boundary**: Brood Box itself is pure Go (`CGO_ENABLED=0`). The embedded go-microvm-runner was pre-built with CGO elsewhere — no CGO needed at bbox build time.
+- **`bbox_full` build tag**: The `bbox-init` guest binary is embedded only under the `bbox_full` build tag (`task build`/`task install` set it alongside `embed_runtime`). Without the tag (the default for `go test`, `golangci-lint`, and library consumers importing `pkg/runtime`/`pkg/sandbox`) `initbin.Binary` is an empty stub, so the module compiles even though the `.gitignore`d `bbox-init` file is absent. A stub-built binary that tries to boot a VM fails fast with an actionable error. If you build `cmd/bbox` by hand, pass `-tags "embed_runtime bbox_full"` or it won't have a working init.
 - **`gh` CLI dependency**: `task fetch-runtime` uses the GitHub CLI (`gh`) to download release artifacts. Firmware is downloaded at runtime via HTTPS by default.
 - **Domain purity**: `pkg/domain/` must never import from `internal/infra/` or `pkg/sandbox/`. This is the most important architectural invariant — break it and you break the entire DDD foundation.
 - **Always use `task`**: Never run `go build`, `go test ./...`, `golangci-lint`, `go fmt`, or `goimports` directly. The Taskfile sets critical env vars and flags. Raw commands will silently produce wrong results.
